requirements.txt
crewai
crewai[tools]
langchain_openai
langchain-community
pydantic
tavily-python
jupyter